fix: termination condition
With the original condition, the loop will not terminate
if there is a configuration error which causes
return codes different from ETIMEDOUT, which is a bug.

There is a more fundamental API issue to the handling
of interrupts in Linux/POSIX: Normally, calls to message
queues are not interrupted like calls to blocking
i/o devices, for example. Interrupting in case
of signals can be activated using the sigaction function().

In the default case, however, the mq functions
will ignore signals.

It might make sense to add an extra flag that enables
early return on ETIMEDOUT.